nmapy/requirements.txt
python-nmap
pyfiglet
pynpyt==1.6.8
